+++ packages/mobx-state-tree/src/types/utility-types/frozen.ts
@@ -11,7 +11,9 @@ import {
     typeCheckFailure,
     TypeFlags,
     isType,
-    ObjectNode
+    ObjectNode,
+    optional,
+    IType
 } from "../../internal"
 
 export class Frozen<T> extends Type<T, T, T> {
@@ -43,6 +45,8 @@ export class Frozen<T> extends Type<T, T, T> {
     }
 }
 
+const frozenInstance = new Frozen()
+
 /**
  * Frozen can be used to story any value that is serializable in itself (that is valid JSON).
  * Frozen values need to be immutable or treated as if immutable. They need be serializable as well.
@@ -52,10 +56,12 @@ export class Frozen<T> extends Type<T, T, T> {
  *
  * Note: if you want to store free-form state that is mutable, or not serializeable, consider using volatile state instead.
  *
+ * `types.frozen` accepts an optional default value. Note that the function is generic to support strong typing
+ *
  * @example
  * const GameCharacter = types.model({
  *   name: string,
- *   location: types.frozen
+ *   location: types.frozen({ x: 0, y: 0})
  * })
  *
  * const hero = GameCharacter.create({
@@ -66,9 +72,20 @@ export class Frozen<T> extends Type<T, T, T> {
  * hero.location = { x: 10, y: 2 } // OK
  * hero.location.x = 7 // Not ok!
  *
+ * @example
+ * type Point = { x: number, y: number }
+ *    const Mouse = types.model({
+ *         loc: types.frozen<Point>()
+ *    })
+ *
  * @alias types.frozen
  */
-export const frozen: ISimpleType<any> = new Frozen()
+export function frozen<T>(defaultValue: T): IType<T | undefined, T, Readonly<T>>
+export function frozen<T>(): IType<T, T, Readonly<T>>
+export function frozen<T>(defaultValue?: any) {
+    if (arguments.length === 0) return frozenInstance as any
+    else return optional(frozenInstance, defaultValue) as any
+}
 
 export function isFrozenType(type: any): type is Frozen<any> {
     return isType(type) && (type.flags & TypeFlags.Frozen) > 0